refactor: Money를 long 자료형으로 변경 및 0으로 초기화 하는 생성자 생성
- LottoResult의 총 상금을 계산하는 메서드 리네이밍

+++ src/main/java/lotto/domain/Lottos.java
@@ -22,12 +22,12 @@ public class Lottos {
     }
 
     public static Lottos from(Money money) {
-        return new Lottos(money.getMoney() / LOTTO_PRICE);
+        return new Lottos((int)(money.getMoney() / LOTTO_PRICE));
     }
 
     public static Lottos of(List<Lotto> lottoList, Money change) {
         Lottos lottos = new Lottos(lottoList);
-        lottos.addAutoCreatedLotto(change.getMoney() / LOTTO_PRICE);
+        lottos.addAutoCreatedLotto((int)(change.getMoney() / LOTTO_PRICE));
         return lottos;
     }
